LayoutTests/platform/glib/imported/w3c/web-platform-tests/loading/preloader-template-expected.txt

main frame - didStartProvisionalLoadForFrame
main frame - didCommitLoadForFrame
main frame - didFinishDocumentLoadForFrame
main frame - didHandleOnloadEventsForFrame

PASS Things inside templates are not preloaded
